% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{aggr.quote}
\alias{aggr.quote}
\title{aggr.quote}
\usage{
aggr.quote(quote, freq, var = "close")
}
\arguments{
\item{quote}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset price}

\item{freq}{An interval specification, one of "day", "week", "month", "quarter" and "year", optionally preceded by an integer and a space, or followed by "s".See \code{\link{cut.Date}} for detail.}

\item{var}{a character string, indicating the aggregated variable,could be:"open","close","high","low","vol",by default "close".}
}
\value{
a aggregated quote series, of the same class of \code{quote}.
}
\description{
aggregate the asset quote series into specific freq.
}
\details{
The function could be used in quote of "open","close","high","low","vol",by specify different \code{var} parametre
}
\examples{
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+1:100)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
price.close <- WealthIndex(rtn)
aggr.quote(price.close,"month")
}
\author{
Ruifei.yin
}
