% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{aggr.rtn}
\alias{aggr.rtn}
\title{aggr.rtn}
\usage{
aggr.rtn(rtn, freq)
}
\arguments{
\item{rtn}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{freq}{An interval specification, one of "day", "week", "month", "quarter" and "year", optionally preceded by an integer and a space, or followed by "s".See \code{\link{cut.Date}} for detail.}
}
\value{
a aggregated return series, of the same class of \code{rtn}.
}
\description{
aggregate the return series into specific freq
}
\examples{
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+1:100)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
aggr.rtn(rtn,"month")
}
\author{
Ruifei.yin
}
