% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{annulized}
\alias{IC.annualized}
\alias{Return.annualized}
\alias{SharpeRatio.annualized}
\alias{StdDev.annualized}
\alias{Table.Annualized}
\alias{Turnover.annualized}
\alias{annulized}
\title{annualized functions}
\usage{
Turnover.annualized(R)

Return.annualized(R, geometric = TRUE)

StdDev.annualized(R)

IC.annualized(R)

SharpeRatio.annualized(R, Rf = 0, geometric = TRUE)

Table.Annualized(R, Rf = 0, geometric = TRUE, digits = 4)
}
\arguments{
\item{R}{a return, stddev, turnover series, an object of class timeSeries,zoo or xts}

\item{geometric}{a logical. generate geometric (TRUE) or simple (FALSE) returns, default TRUE}
}
\value{
a vector or scalar depending on the dim of seri
}
\description{
calculate an annualized return, stddev, turnover.
}
\examples{
#-- turnover.annulized
seri <- zoo(runif(30,0,1),seq(Sys.Date(),by="month",length.out=30))
re <- Turnover.annualized(seri)
seri <- zoo(matrix(runif(60,0,1),30,2),seq(Sys.Date(),by="month",length.out=30))
re <- Turnover.annualized(seri)
#-- return.annulized
#- monthly
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+(1:100)*30)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
Return.annualized(rtn) 
PerformanceAnalytics::Return.annualized(rtn) 
#- 10day
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+(1:100)*10)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
Return.annualized(rtn) # right
PerformanceAnalytics::Return.annualized(rtn) # wrong!
}
\author{
Ruifei.Yin
}

