% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{check.name_exist}
\alias{check.name_exist}
\title{check.name_exist}
\usage{
check.name_exist(obj)
}
\arguments{
\item{obj}{any object to be checked}
}
\description{
check if a object have names
}
\author{
Ruifei.Yin
}
