% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{findInterval.rightClosed}
\alias{findInterval.rightClosed}
\title{findInterval.rightClosed}
\usage{
findInterval.rightClosed(x, vec, ...)
}
\description{
findInterval.rightClosed
}
\seealso{
\code{\link{findInterval}}
}
