% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{getBiCop}
\alias{getBiCop}
\title{Get bivariate random variables}
\usage{
getBiCop(n, rho, mar.fun = rnorm, x = NULL, drop.x = FALSE, ...)
}
\arguments{
\item{n}{a integer. number of observations}

\item{rho}{a numeric. the desired correlation}

\item{mar.fun}{the marginal distribution function}

\item{x}{a vector. the existing variable.}

\item{drop.x}{logical. If true, then return a vector of the generated variable.}

\item{...}{other arguments to mar.fun}
}
\value{
a data frame of two variables which correlate with a population correlation of rho, or a vector when drop.x is true.
}
\description{
Compute a bivariate sample distribution drawn from a population with a given . It either computes two random variables, or it takes one existing variable (passed as parameter x) and creates a second variable with the desired correlation.
}
\details{
Get from a forum post: \url{http://stats.stackexchange.com/questions/15011/generate-a-random-variable-with-a-defined-correlation-to-an-existing-variable/15035#15035}
}
\note{
this is only an approximate solution, i.e., the empirical correlation is not exactly equal to rho.
}
\examples{
# -- generate two random variables
df <- getBiCop(100,0.2)
cor(df)
# -- generate another variable with desired correlation on a existing variable
x <- rnorm(100)
df2 <- getBiCop(100,0.2,x=x)
cor(df2)
}
