% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{getEventRtn}
\alias{getEventRtn}
\alias{getEventRtns}
\alias{getEventRtns}
\title{getEventRtn}
\usage{
getEventRtn(TS, holdingday, bmk = "EI000300", fee = 0,
  tradeType = c("close", "nextavg", "nextopen"))

getEventRtns(TS, holdingdays, bmk = "EI000300", fee = 0,
  tradeType = c("close", "nextavg", "nextopen"))
}
\arguments{
\item{TS}{a TS object.}

\item{holdingday}{a integer, giving the holding day.}

\item{bmk}{a character string,giving the stockID of the benchmark index,eg. "EI000300".}

\item{fee}{a mumeric, giving the one side fee of trading.}

\item{tradeType}{a character string("close","nextavg","nextopen"),indicating the trading type.See detail in \code{\link{getPeriodrtn}}}

\item{holdingdays}{a vector of integers,giving the list of holding days}
}
\value{
getEventRtn return a data frame with cols:"stockID","date","endT","rtn","bmkRtn","exRtn".

getEventRtns return a data frame(with cols same as return of getEventRtn),which is a rbind of returns of getEventRtn by different holdingdays.
}
\description{
Given a event data(ussually a TS object), calculate the extra returns driven by the event.
}
\examples{
eventRtn <- getEventRtn(TS,20)
eventRtns <- getEventRtns(head(TS,10),c(-10,-5,0,5,10))
}
\author{
Ruifei.Yin
}
