% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{ggplot.Drawdown}
\alias{ggplot.Drawdown}
\title{ggplot.Drawdown}
\usage{
\method{ggplot}{Drawdown}(rtn, geometric = TRUE, main = NULL, alpha = if
  (dim(rtn)[2] > 1) 0.5 else 1, ...)
}
\description{
ggplot.Drawdown
}
