% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{ggplot.ts.area}
\alias{ggplot.ts.area}
\title{ggplot.ts.area}
\usage{
\method{ggplot}{ts.area}(ts, main = NULL, ...)
}
\arguments{
\item{ts}{a zoo, timeSeries or xts object}

\item{main}{chart title , character string}

\item{...}{other arguments passed to function \code{geom_area}}
}
\value{
a area chart of class \code{ggplot}
}
\description{
ggplot.ts.area
}
\examples{
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+1:100)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
ggplot.ts.area(rtn)
}
\author{
Ruifei.Yin
}
