% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{ggplot.ts.bar}
\alias{ggplot.ts.bar}
\title{ggplot.ts.bar}
\usage{
\method{ggplot}{ts.bar}(ts, main = NULL, ...)
}
\arguments{
\item{ts}{a zoo, timeSeries or xts object}

\item{main}{}

\item{...}{other arguments passed to function \code{geom_bar}}
}
\value{
a bar chart of cloass \code{ggplot}
}
\description{
ggplot.ts.bar
}
\examples{
rtn.long <- zoo(rnorm(10,0.001,0.02),as.Date("2010-01-01")+1:10)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
ggplot.ts.bar(rtn)
}
\author{
Ruifei.Yin
}
