% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{ggplot.ts.line}
\alias{ggplot.ts.line}
\title{ggplot.ts.line}
\usage{
\method{ggplot}{ts.line}(ts, facet = FALSE, main = NULL, ...)
}
\arguments{
\item{ts}{a zoo, timeSeries or xts object}

\item{main}{chart title , character string}

\item{...}{other arguments passed to function \code{geom_line}}
}
\value{
a line chart of class \code{ggplot}
}
\description{
ggplot.ts.line
}
\examples{
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+1:100)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
ggplot.ts.line(rtn)
}
\author{
Ruifei.Yin
}
