% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{ggplot.WealthIndex}
\alias{ggplot.WealthIndex}
\title{ggplot.WealthIndex}
\usage{
\method{ggplot}{WealthIndex}(rtn, geometric = TRUE, facet = FALSE,
  main = NULL, ...)
}
\description{
ggplot.WealthIndex
}
