% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{ggplots.RollingPerformance}
\alias{ggplots.RollingPerformance}
\title{ggplots.RollingPerformance}
\usage{
ggplots.RollingPerformance(rtn, width = 250, by = 30, ...)
}
\arguments{
\item{rtn}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{width}{see \code{\link[zoo]{rollapply}} for detail}

\item{by}{see \code{\link[zoo]{rollapply}} for detail}

\item{...}{other arguments to function \code{\link{rollingPerformance}}}

\item{align}{see \code{\link[zoo]{rollapply}} for detail}
}
\value{
Print a wrapped plots of class ggplot,and return a recordedplot object.
}
\description{
A wrapper to create a rolling annualized returns chart(of class ggplot), rolling annualized standard deviation chart, and a rolling annualized sharpe ratio chart.
}
\examples{
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+1:100)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
ggplots.RollingPerformance(rtn)
}
\seealso{
\code{\link{rollingPerformance}}
}
\author{
Ruifei.Yin
}
