% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{hitRatio}
\alias{hitRatio}
\title{hitRatio}
\usage{
hitRatio(rtn, satisfied = 0)
}
\arguments{
\item{rtn}{a vector or a matrix(or a matrix-like object,eg.xts,zoo,timeSeries,..each column gives a rtn series of an asset)}

\item{satisfied}{a numeric, indicating how much return could be called a "hit"}
}
\value{
the hit ratio, a matrix with dims 1*NCOL(rtn)
}
\description{
get the hit ratio of the return series.
}
\examples{
rtn <- matrix(runif(100,0,0.05),50,2)
hit <- hitRatio(rtn,0.02)
}
\author{
Ruifei.Yin
}
