% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{html.characters}
\alias{html.characters}
\title{html.characters}
\usage{
html.characters(characters, file = get(".HTML.file"))
}
\arguments{
\item{characters}{A vector of character string}

\item{file}{A charactor string, giving the path and the file name of the html report file}
}
\description{
Write characters to the HTML file by writing one or more <p> tags
}
\details{
This function is a wraped function of the \code{\link[R2HTML]{HTML.character}} function in the package \code{R2HTML}. The differenc is that the \code{HTML.character} write a character vector to one paragraph,while the \code{html.characters} to numbers(same with the length of the vector) of paragraphs
}
\examples{
.HTML.file <-  file.path(getwd(),"htmlrpt.html")
html.characters(c("this is the first paragraph.","this is the second paragraph."))
}
\seealso{
\code{\link[R2HTML]{HTMLInsertGraph}},\code{\link[R2HTML]{HTMLplot}}
}
\author{
Ruifei.Yin
}
