% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{html.initfile}
\alias{html.initfile}
\title{html.initfile}
\usage{
html.initfile(file = get(".HTML.file"), Title = "R output",
  CSSFile = "R2HTML.css", HTMLframe = FALSE, BackGroundColor = "FFFFFF",
  BackGroundImg = "", useLaTeX = TRUE, useGrid = TRUE)
}
\arguments{
\item{file}{A charactor string, giving the path and the file name of the HTML file}

\item{Title}{The title of the HTML file,the string to pass to HTML <title> tag}

\item{CSSFile}{Name of a CSS file to use,or absolute path and name when the the css file is not in the same directory with the target HTML file}

\item{...}{More arguments to the \code{\link[R2HTML]{HTMLInitFile}} function in the package \code{R2HTML}}
}
\description{
Initiate a HTML file by writing the HTML <body><head><title></title></head>...</body> tags and their options.
}
\details{
This function is a wraped function(but more convenient to use) of the \code{\link[R2HTML]{HTMLInitFile}} function in the package \code{R2HTML}
}
\examples{
htmlrptDir <- file.path(getwd(),"htmlrpt")
htmlrtpFile <- "single_factor_model_report.html"
cssFile <- "factormodelrpt.css"
cssFile <- paste("file:///",file.path(htmlrptDir,cssFile),sep="") # if the css file is not in the same directory with the target HTML file, a absolute path is needed
.HTML.file <- file.path(htmlrptDir,htmlrtpFile)
html.initfile(CSSFile=cssFile,Title="Factor model report")
}
\seealso{
\code{\link[R2HTML]{HTMLInitFile}}
}
\author{
Ruifei.Yin
}
