% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{html.plot}
\alias{html.plot}
\title{html.plot}
\usage{
html.plot(plotObj, file = get(".HTML.file"), Caption = "",
  GraphBorder = 0, Align = "center", Width = 500, Height = 500,
  WidthHTML = NULL, HeightHTML = NULL, GraphPointSize = 12,
  GraphBackGround = "white")
}
\arguments{
\item{plotObj}{A plot object which could be the result of a plot function or a ggplot function.}

\item{file}{A charactor string, giving the path and the file name of the html report file}

\item{Caption}{If non empty, text to be written under the graph, as its caption}

\item{...}{More arguments to the \code{\link{png}} function and \code{\link[R2HTML]{HTMLInsertGraph}} function in the package \code{R2HTML}}
}
\description{
Exports the active graphic to a \code{png} file and add it to a target HTML output, by writing the <IMG> tag.
}
\details{
This function is a wraped function of the \code{\link[R2HTML]{HTMLInsertGraph}} function in the package \code{R2HTML}
}
\examples{
.HTML.file <-  file.path(getwd(),"htmlrpt.html")
html.plot(plot(1:5))
}
\seealso{
\code{\link[R2HTML]{HTMLInsertGraph}},\code{\link[R2HTML]{HTMLplot}}
}
\author{
Ruifei.Yin
}
