% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{lag.m}
\alias{lag.m}
\title{lag.m}
\usage{
lag.m(m, k = 1, na.pad = FALSE)
}
\arguments{
\item{m}{a matrix or a vector}

\item{k}{the number of lags (in units of observations).Note the sign of k: a series lagged by a positive k is shifted earlier,and a negtive k is shifted later.}

\item{na.pad}{logical. If TRUE it adds any rows that would not otherwise have been in the result with a value of NA. If FALSE those rows are dropped.}
}
\value{
a matirx or a vector
}
\description{
Methods for computing lags of matrix or vector objects.
}
\examples{
m <- matrix(1:10,5,2)
rownames(m) <- letters[1:5]
colnames(m) <- LETTERS[1:2]
lag.m(m,1)
lag.m(m,-1)
lag.m(m,1,TRUE)
v <- c(a=1,b=2,c=3,d=4)
lag.m(v,-2,TRUE)
}
\author{
Ruifei.yin
}
