% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{ls.class}
\alias{ls.class}
\title{ls.class}
\usage{
ls.class(class, pattern = "")
}
\arguments{
\item{class}{:character string or a character vector, giving the class of objects to list}

\item{pattern}{:an optional regular expression.}
}
\value{
a character vector,giving the names of the objects in the global environment with the given classes
}
\description{
list the names of objects with certain class
}
\examples{
ls.class(c("function","character"))
}
\author{
Ruifei.Yin
}
