% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{melt.ts}
\alias{melt.ts}
\title{melt.ts}
\usage{
\method{melt}{ts}(ts)
}
\arguments{
\item{ts}{a zoo, timeSeries or xts object}
}
\value{
a data frame object melted from the ts,with cols:time,variable,value
}
\description{
melt the multi-columns timeseries to single-column dataframe for easy ggploting.
}
\examples{
rtn.long <- zoo(rnorm(10,0.001,0.02),as.Date("2010-01-01")+1:10)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
df <- melt.ts(rtn)
}
\author{
Ruifei.Yin
}
