% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{merge.x}
\alias{merge.x}
\title{merge.x}
\usage{
\method{merge}{x}(x, y, by = intersect(colnames(x), colnames(y)),
  mult = "all")
}
\arguments{
\item{x}{a data frame}

\item{y}{a data frame}

\item{by}{a vector of shared column names in x and y to merge on.}

\item{mult}{a character string: one of "all","first","last". See detail in \code{\link[data.table]{data.table}}}
}
\value{
a new data.frame based on the merged data tables, sorted by the x order.
}
\description{
Just same with \code{merge}, except that: (1) the order of merging result is kept same with \code{x};(2)speed up by using \code{merge.data.table}.
}
\examples{
x <- data.frame(k1=c(1,5,3),k2=c("a","a","c"),V=11:13)
y <- data.frame(k1=c(2,3,1,5,6),k2=c("b","c","a","f","x"),v1=LETTERS[1:5],v2=LETTERS[8:12])
# compare the following result:
merge.x(x,y)
merge(x,y)
merge(x,y,all.x=TRUE)
dplyr::left_join(x,y)
}
\author{
Ruifei.Yin
}
