% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{multiplot_facet}
\alias{multiplot_facet}
\title{multiplot_facet}
\usage{
multiplot_facet(plotlist, nrow = NULL, ncol = NULL, scales = "fixed",
  shrink = TRUE, as.table = TRUE, drop = TRUE)
}
\arguments{
\item{plotlist}{A list of ggplot objects}

\item{nrow}{Number of rows of plots}

\item{ncol}{Number of columns of plots}
}
\value{
a ggplot object, with factets of elements of plotlist. The naming rules of the facet label is: if plotlist have names, then use them; else if the element of plotlist have titles, then use them; else, use "plot_1" to "plot_n".
}
\description{
put multiple ggplot objects on one page with facets.
}
\examples{
p1 <- qplot(y=1:4)
p2 <- qplot(1:5,10:14)
p3 <- qplot(4:30)
p4 <- qplot(3:9,4:10)
multiplot_facet(list(p1, p2, p3, p4), ncol=2, scales="free")
}
\author{
Ruifei.Yin
}
