% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{periodicity_Ndays}
\alias{periodicity_Ndays}
\title{periodicity_Ndays}
\usage{
periodicity_Ndays(x)
}
\arguments{
\item{x}{a timeseries object or a Date vector}
}
\value{
a numeric
}
\description{
Estimate the periodicity of a time-series-like object by calculating the mean time between observations in days.
}
\seealso{
\link[xts]{periodicity}
}
