% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{renameCol}
\alias{renameCol}
\title{Rename columns in a matrix or a dataframe.}
\usage{
renameCol(indata, src, tgt)
}
\arguments{
\item{indata}{A dataframe or a matrix}

\item{src}{Source: Vector of names of columns in 'indata' to be renamed. Can also be a vector of column numbers.}

\item{tgt}{Target: Vector with corresponding new names in the output.}
}
\value{
A dataframe if 'indata' is a dataframe; a matrix in 'indata' is a matrix.
}
\description{
Rename columns in a matrix or a dataframe.
}
\note{
This function is imported from \code{\link[doBy]{renameCol}}.
}
\examples{
CO2 <- data.frame(a=1,b=2)
renameCol(CO2, 1:2, c("kk","ll"))
renameCol(CO2, c("a","b"), c("kk","ll"))
}
