% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{Returns}
\alias{Returns}
\title{Returns}
\usage{
Returns(prices, geometric = TRUE, na.rm = TRUE, trim = TRUE)
}
\arguments{
\item{prices}{price or index series, an object of class timeSeries,zoo or xts}

\item{geometric}{use geometric (TRUE) or simple (FALSE) returns, default TRUE}

\item{na.rm}{a logical value. Should NAs be removed? By Default TRUE.}

\item{trim}{a logical value. Should the time series be trimmed? By Default TRUE.}
}
\value{
the returns series of the same class of \code{prices}
}
\description{
Compute financial returns from prices or indexes.
}
\details{
This is a wrapped function of function \code{\link[timeSeries]{returns}} in package \code{timeSeries} which is used for timeSeries objects
}
\examples{
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+1:100)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
prices <- WealthIndex(rtn)
rtn2 <- Returns(prices,trim=FALSE) # Is rtn2 the same as rtn?
}
\seealso{
\code{\link{WealthIndex}},\code{\link[timeSeries]{returns}}
}
\author{
Ruifei.Yin
}
