% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{rollingPerformance}
\alias{rollingPerformance}
\title{rollingPerformance}
\usage{
rollingPerformance(rtn, FUN = "Return.annualized", width = 365, by = 30,
  align = "right", ...)
}
\arguments{
\item{rtn}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{FUN}{any function that can be evaluated using a single set of returns (e.g., rolling CAPM.beta won't work, but Return.annualized will)}

\item{width}{see \code{\link[zoo]{rollapply}} for detail}

\item{by}{see \code{\link[zoo]{rollapply}} for detail}

\item{align}{see \code{\link[zoo]{rollapply}} for detail}

\item{...}{other arguments to function FUN or \code{\link[zoo]{rollapply}}}
}
\value{
A object of class xts with the results of the rolling function
}
\description{
get the rolling performance of the return series
}
\examples{
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+1:100)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
re <- rollingPerformance(rtn, width=20, by=5)
}
\author{
Ruifei.Yin
}
