% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{rptDate.yoy}
\alias{rptDate.yoy}
\title{rptDate.yoy}
\usage{
rptDate.yoy(rptDate)
}
\arguments{
\item{rptDate}{a vector of rptDate , with class Date.}
}
\value{
a vector of yoy rptDate.
}
\description{
rptDate.yoy
}
\examples{
rptDate.yoy(ymd(c(20100630,20141231,20130930)))
}
\author{
Ruifei.Yin
}

