% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{rtn.lastperiods}
\alias{rtn.lastperiods}
\title{rtn.lastperiods}
\usage{
rtn.lastperiods(rtn, periods = list(months(1), months(3), months(6), years(1),
  years(3), years(5)))
}
\arguments{
\item{rtn}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{periods}{a list, with elements of class "Period",eg.list(months(1),months(3)), giving the last periods.}
}
\value{
a matrix
}
\description{
get a return of different last periods of the rtn series(including overall cumulativereturn).
}
\examples{
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+1:100)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
rtn.lastperiods(rtn,list(months(3),months(6),years(1)))
}
\author{
Ruifei.Yin
}
