% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{rtn.periods}
\alias{rtn.periods}
\title{rtn.periods}
\usage{
rtn.periods(rtn, freq = "year", from, to)
}
\arguments{
\item{rtn}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{freq}{An interval specification, one of "day", "week", "month", "quarter" and "year", optionally preceded by an integer and a space, or followed by "s".See \code{\link{cut.Date}} for detail.}

\item{from}{starting dates, a vector of class Date,timeDate or character.See \code{\link[timeSeries]{fapply}} for detail.}

\item{to}{ending dates  a vector of class Date,timeDate or character.See \code{\link[timeSeries]{fapply}} for detail.}
}
\value{
a matrix
}
\description{
get a return of different periods of the rtn series(including overall cumulative and annualized return).
}
\examples{
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+1:100)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
rtn.periods(rtn,"month")
rtn.periods(rtn,from=c("2010-02-03","2010-03-04"),to=c("2010-03-09","2010-03-30"))
}
\seealso{
\code{\link[timeSeries]{fapply}}
}
\author{
Ruifei.Yin
}
