% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{Turnover.annualized}
\alias{Turnover.annualized}
\title{Turnover.annualized}
\usage{
Turnover.annualized(seri)
}
\arguments{
\item{seri}{a turnover series,an object of class timeSeries,zoo or xts}
}
\value{
a vector or scalar depending on the dim of seri
}
\description{
calculate an annualized turnover rate from a turnover series
}
\examples{
seri <- zoo(runif(30,0,1),seq(Sys.Date(),by="month",length.out=30))
re <- Turnover.annualized(seri)
seri <- zoo(matrix(runif(60,0,1),30,2),seq(Sys.Date(),by="month",length.out=30))
re <- Turnover.annualized(seri)
}
\author{
Ruifei.Yin
}

