% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{vlookup.df}
\alias{vlookup.df}
\title{vloolup.df}
\usage{
vlookup.df(x, table, by = colnames(x), ret = setdiff(colnames(table), by),
  exclude.x = TRUE)
}
\arguments{
\item{x}{a data frame}

\item{table}{a data frame}

\item{by}{a vector of character string}

\item{ret}{a vector of character string or integer}

\item{exclude.x}{a logical. If the \code{x} shoud be inculded in the result?}
}
\value{
a data frame (or a vector if \code{exclude.x} is TRUE and length of \code{ret} is 1).
}
\description{
vloolup.df
}
\examples{
x <- data.frame(m1=1:3,m2=c("a","b","c"))
table <- data.frame(K1=c(2,3,1,5,6),K2=c("b","c","a","f","x"),v1=LETTERS[1:5],v2=LETTERS[8:12])
vlookup.df(x,table,c("K1","K2"))

y <- data.frame(m1=c(1,3,1),m2=c("a","c","a"))
vlookup.df(y,table,c("K1","K2"))

z <- data.frame(m1=c(1,5,3),m2=c("a","a","c"))
vlookup.df(z,table,c("K1","K2"))
vlookup.df(z,table,c("K1","K2"), exclude.x=FALSE)
}
