% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{vlookup}
\alias{vlookup}
\title{vlookup}
\usage{
vlookup(x, table, by, ret)
}
\arguments{
\item{x}{a vector}

\item{table}{a matrix, data frame or other matrix like object}

\item{by}{a vector of column names or column index}

\item{ret}{a vector of column names or column index}
}
\value{
a matrix, data frame or vector, corresponding with the class of \code{table} and length of \code{ret}.
}
\description{
vlookup
}
\examples{
x <- c(2,4,3)
table <- data.frame(1:4,LETTERS[1:4])
vlookup(x,table,by=1,ret=2)
}
