% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{WealthIndex}
\alias{WealthIndex}
\title{WealthIndex}
\usage{
WealthIndex(rtn, geometric = TRUE)
}
\arguments{
\item{rtn}{return series, an object of class timeSeries,zoo or xts}

\item{geometric}{get geometric (TRUE) or simple (FALSE) returns, default TRUE}
}
\value{
the wealth index series of the same class of \code{rtn}
}
\description{
Compute cumulated financial series, e.g. prices or indexes, from financial returns.
}
\details{
This is a wrapped function of function \code{\link[timeSeries]{cumulated}} in package \code{timeSeries} which is used for timeSeries objects
}
\examples{
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+1:100)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
prices <- WealthIndex(rtn)
}
\seealso{
\code{\link{Returns}},\code{\link[timeSeries]{cumulated}}
}
\author{
Ruifei.Yin
}
