% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{check.colnames}
\alias{check.colnames}
\alias{usualcols}
\alias{is_usualcols}
\alias{guess_factorNames}
\alias{check.TS}
\alias{check.rptTS}
\alias{check.TSF}
\alias{check.TSFR}
\alias{check.Port}
\alias{check.SP}
\alias{check.TSS}
\title{check.colnames}
\usage{
check.colnames(data, coltest)

usualcols()

is_usualcols(cols)

guess_factorNames(df, no_factorname = NULL, is_factorname = "factorscore",
  silence = FALSE)

check.TS(TS)

check.rptTS(rptTS)

check.TSF(TSF)

check.TSFR(TSFR)

check.Port(port)

check.SP(SP)

check.TSS(TSS)
}
\arguments{
\item{data}{a dataframe or matrix,with some colnames}

\item{coltest}{a character vector containing the colnames to be checked}
}
\description{
check colnames of an object with data.frame class, eg. TS, TSF, TSFR, ... Return error when the elements of coltest is not contained in the colnames of data
}
\examples{
df <- data.frame(a=1,b=2)
check.colnames(df,c("a","b"))
check.colnames(df,c("a","c"))
df <- data.frame(a=1,b=2,c=3,date=9,factorscore=0)
guess_factorNames(df,no_factorname = "c",is_factorname = "factorscore")
}
\author{
Ruifei.Yin
}
