% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{ggplot.corr}
\alias{ggplot.corr}
\title{ggplot.corplot}
\usage{
\method{ggplot}{corr}(corr)
}
\arguments{
\item{corr}{is correlation matrix or a list of correlation matrix ,see \code{\link[stats]{cor}}.}
}
\description{
ggplot.corplot
}
