% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pub01_utilityFuncs.R
\name{rtn.summary}
\alias{rtn.summary}
\title{rtn.summary}
\usage{
rtn.summary(rtn, hitFreq = "day", hitSatisfied = 0, Rf = 0)
}
\arguments{
\item{rtn}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{hitFreq, indicating}{the interval when computing the hitRatio of rtn. An interval specification, one of "day", "week", "month", "quarter" and "year", optionally preceded by an integer and a space, or followed by "s".See \code{\link{cut.Date}} for detail.}

\item{hitSatisfied}{a numeric, indicating how much return could be called a "hit".See \code{\link{hitRatio}} for detail.}

\item{Rf}{risk free rate, in same period as your returns}
}
\value{
a matrix, giving the summary infomation of the rtn series,including Annualized Return,Annualized Std Dev,Annualized Sharpe,HitRatio,Worst Drawdown
}
\description{
get the summary infomation of the rtn series,including Annualized Return,Annualized Std Dev,Annualized Sharpe,HitRatio,Worst Drawdown
}
\examples{
rtn.long <- zoo(rnorm(100,0.001,0.02),as.Date("2010-01-01")+1:100)
rtn.short <- rtn.long + rnorm(100,-0.001,0.003)
rtn <- merge(rtn.long,rtn.short)
rtn.summary(rtn)
}
\author{
Ruifei.Yin
}
